backend/requirements.txt
celery==5.3.6
fastapi==0.111.0
nltk==3.8.1
pydantic==1.10.11
pymongo==3.13.0
pypdf==4.2.0
spacy==3.7.5
stanford_openie==1.3.2
redis==5.0.3
